reorganize generator
